+++ src/rhs.c
@@ -149,7 +149,7 @@ typedef struct {
 static RhsGraphic rhs_graphic[] = {
     // { CLASS_CLOCK, clock_get_type },
     { CLASS_EXPRESSION, expression_get_type },
-    // { CLASS_GROUP, group_get_type },
+    { CLASS_GROUP, group_get_type },
     // { CLASS_LIST, group_get_type },
     { CLASS_PATHNAME, pathname_get_type },
     { CLASS_FONTNAME, fontname_get_type },
@@ -163,8 +163,8 @@ static RhsGraphic rhs_graphic[] = {
     { CLASS_PLOT, plot_get_type },
     { CLASS_IMAGE, iimage_get_type },
     { CLASS_NUMBER, number_get_type },
-    // { CLASS_REAL, real_get_type },
-    // { CLASS_VECTOR, vector_get_type },
+    { CLASS_REAL, real_get_type },
+    { CLASS_VECTOR, vector_get_type },
     { CLASS_STRING, string_get_type }
 };