+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - add ^G to group selected rows
 - add "Group selected" to column menu
 - toolkitgroupview is more compact
+- add vips_header_get builtin
+- add Real, Group, Vector graphic classes
 
 ## 9.0.11 2025/07/21